refactor: Consolidate data and process controllers into AssetController
- Created AssetController with clear asset management responsibilities
- Consolidated functionality from DataController and ProcessController
- Improved file handling and processing logic
- Enhanced error handling with detailed logging
- Added support for different file types and processing strategies
- Provides clear separation of concerns for asset operations
